Stops throwing exceptions with __toString() message
Desc:
In PHP exceptions thrown from __toString() function does not provide
a proper stack trace and are very hard to debug. This is currently
resulting in a lot of error exceptions where emails are sent and
there may be some minor issue in converting the laravel views with
attached images using the __toString function. Side-effect of the issue
is that we are getting lots of useless alerts on which we cannot take
any effective action. This issue has finally been fixed in PHP 7.0
Till the migration to 7 happens, we are going to ignore the exceptions
throw because of this issue.