[es] Pushes entity array instead of entity into es store queue.
Desc:
* After laravel upgrade, now Eloquent entity being pushed into queue
is automatically converted to an id which is then fetched from database
upon resolving during the job being processed. This causes issues for us
because we have two databases that we juggle with and hence we cannot
push eloquent models as payload in queues anymore. Instead it's being
changed to array in for es queue."