[OAuth] Adding bearer auth tests (#4677)
* [OAuth] Adding bearer auth tests

* unset token from trace data

* Moving to token fixture

* Adding more test cases

* Adding prod client test

* Fixing conflict errors

* [oauth] Following review fixes:

- Updates composer.lock to latest dev-master of oauth package, a fix there
- Renames method in tests/Authorization file
- Moves new tests to OAuth namespace
- Fixes OAuthServer instantiation

* [oauth] Minor fixes in handler

* [oauth] Default oauth rsa pair in .wercker

There are 2 tests in API which runs in environtment=production.
OAuthServer() looks for RZP_OAUTH_{PUBLIC|PRIVATE}_KEY in .env.production -> .env.defaults. And couldn't find anything.
For easy dev/testing those keys are committed in oauth package (in /keys folder) but that only gets picked when environtment
is testing. So basically for now adding above defaults in .env.wercker so in any environment that gets picked.

Not adding the same in .env.defaults because if the same env variables are not set in .env.production then those
will work on production as well which can be a security issue.

We are having those env values in .env.production.

* Removing extra code from test

* Revert "[oauth] Default oauth rsa pair in .wercker"

This reverts commit 818d1043ef4e074e089e763b562c22e9e00a18e4.

 Also fixes bad merge in prev commit.

 Conflicts:
	environment/.env.wercker

* [oauth] Handles getallheaders() issues in tests

- Handles getallheaders() not working in unit tests; Refer comments
- Minor other fixes in OAuth's test traits