Deps to be ready for integration via twirp client (#17300)
* Adds deps- guzzlehttp/psr7 and http-interop/http-factory-guzzle

guzzlehttp/psr7 provides psr/http-factory-implementation. More
specifically, it has psr-17 request and response factories
implementations.

http-interop/http-factory-guzzle is required because of https://github.com/php-http/discovery/issues/164#issuecomment-596015643.

Some other lib can be used but sticking to guzzle as much as
possible, for which there is no specific reason but just that
that is what we have been using.

Also there exists a polyfill func getallheaders in lib/utility.php
which will now be overridden by deps ralouphie/getallheaders. It is compatible.

* Adds kriswallsmith/Buzz

kriswallsmith/Buzz provides psr/http-client-implementation.
It has psr-18 http client factory implementation.

Ideally the deps php-http/guzzle6-adapter itself be upgraded
to ^2 but doing this requries lot of deps hell resolution. It will
require to file and wait on external libs e.g. php-http/laravel-httpplug.
There is possibility of bit more discussions in this approach and want to
do it in free time.

And hence adding this alternative implementation instead. Have verified
that it does not affect existing discovery, and that only for psr-18 discovery
this lib is used.