[RTB] Refactor isNewCustomer Query to Use Invisible Index (#27748)
* testing commit

* Refactor isNewCustomer Query to Use Invisible Index

__Problem:__
The `isNewCustomer` query is taking up to 3 seconds to execute on the reporting replica DB. This will start causing hindrance to customer experience as this query is being used in the `/v1/preferences` endpoint which is used to load checkout.

__Cause:__

The reporting replica has a composite index on merchantId, contact, created_at on `payments` table called `payments_merchant_id_contact_created_at_index_all_replicas`. But this index is disabled i.e. made invisible to the MySQL query optimizer. So, the query optimizer is using the index on `merchant_id` to perform the search.

__Fix:__

We are enabling a optimizer switch `use_invisible_indexes=on` to make this index visible for this SQL query. Also, we have narrowed down the number of columns fetched from `*` to `id`.

* Use Payment Fetch Replica Connection & Set Max Execution Time to 1.5 seconds

Co-authored-by: saikiran <sai.saindla@razorpay.com>
Co-authored-by: rzpcibot <64553331+rzpcibot@users.noreply.github.com>