CE-6104: Fix Tokens Being Returned as HashMap in Preferences & OTP Verify (#31273)
__PROBLEM:__

The `customer.tokens` in `/preferences` and `tokens.items` in `/otp/verify`
fields were being returned as a map instead of an array.

This was caught/reported on QA automation environment.

__CAUSE:__

The newly added method `removeNonActiveTokenisedCardTokens()` was removing a few
entries from the tokens items using `array_filter`. The array_filter method
preserves the keys i.e. the integer keys now have gaps in them. When such array
is converted to JSON then it get's converted to a HashMap as the keys have gaps
& the array isn't general indexed array anymore.

__FIX:__

Add call to `values()` before returning tokens from `removeNonActiveTokenisedCardTokens()`

Co-authored-by: rzpcibot <64553331+rzpcibot@users.noreply.github.com>