CE-7095: BugFix to Ensure QrCode Pricing Is Not Picked for UPI Payments on Checkout During Manual Capture (#34540)
* CE-7095: BugFix to Ensure QrCode Pricing Is Not Picked for UPI Payments on Checkout During Manual Capture

__PROBLEM:__

Upon testing on production we figured out that the issue was only solved for
auto-captured payments. We observed that `defaultQRPricing` was still being
applied for manual captured payments.

__CAUSE:__

The `isUpiQr()` method in `Payment\Entity` makes a call to `isFlowIntent()` which
checks for `flow` in `$this->metadata`. In manual capture mode this metadata
isn't available & is an empty array. Hence the `isFlowIntent()` return false & the
fix put in PR #34464 to make `$receiverType` as `null` doesn't work.

__FIX:__

Refactor `Payment\Entity::isFlowIntent()` method to check for `flow` in UpiMetaData
too as the `flow` information is stored in `upi_metadata` table after the payment
is authorised.

* CE-7095: Check Receiver Type in isQrV2Payment() & Switch to Using Only That in Pricing Calculator

__PROBLEM:__

A lot of UPI Mandate TestCases were failing as the `$this->upiMetadata` was being
loaded before it being created & was ending up becoming null.

__CAUSE:__

The check in `isFlowIntent()` was loading it very early in the payment processing
& that caused the `$upiMetadata` relationship to get cached in the payment entity
with a null value.

__FIX:__

1. Revert `isFlowIntent()` to its original state.
2. Add `receiver_type = 'qr_code'` check in `isQrV2Payment()`
3. Remove usage of `isUpiQr()` in Payment Fee Calculator & switch to
   `isQrV2UpiPayment()`

Co-authored-by: rzpcibot <64553331+rzpcibot@users.noreply.github.com>