ci(.npmrc): remove loglevel from npmrc (#11757)
Setting the default loglevel as silent in the root level npmrc is misleading to users, specially
when npm install fails for some reason. The default setting of loglevel reports errors when they
occur, which is expected behavior from most developers. If logs are to be removed for ci, the flag
should be explicitly set where the command is run, not at the root level npmrc since this confuses
developers working locally.